+++ samples/sampler/stories/default/Card.js
@@ -1,7 +1,8 @@
 import {Card, CardBase} from '@enact/limestone/Card';
+import icons from '@enact/limestone/Icon/IconList';
 import {mergeComponentMetadata} from '@enact/storybook-utils';
 import {action} from '@enact/storybook-utils/addons/actions';
-import {boolean, object, select, text} from '@enact/storybook-utils/addons/controls';
+import {boolean, number, object, select, text} from '@enact/storybook-utils/addons/controls';
 import {Card as UiCard} from '@enact/ui/Card';
 
 import {svgGenerator} from '../helper/svg';
@@ -16,6 +17,8 @@ const generateImageSrc = (color) => {
 		uhd: svgGenerator(600, 600, color, 'ffffff', '600 X 600')
 	};
 };
+
+const iconsList = Object.keys(icons).sort();
 
 const prop = {
 	orientation: ['horizontal', 'vertical']
@@ -33,6 +36,7 @@ export const _Card = (args) => (
 		centered={args['centered']}
 		disabled={args['disabled']}
 		fitImage={args['fitImage']}
+		icon={args['icon']}
 		imageIconSrc={args['imageIconSrc']}
 		imageSize={args['imageSize']}
 		hasContainer={args['hasContainer']}
@@ -41,11 +45,13 @@ export const _Card = (args) => (
 		onClick={action('onClick')}
 		orientation={args['orientation']}
 		primaryBadgeSrc={args['primaryBadgeSrc']}
+		progress={args['progress']}
 		roundedImage={args['roundedImage']}
 		secondaryBadgeSrc={args['secondaryBadgeSrc']}
 		// eslint-disable-next-line no-undefined
 		secondaryLabel={args['secondaryLabel'] ? args['secondaryLabel'] : undefined}
 		selected={args['selected']}
+		showProgressBar={args['showProgressBar']}
 		src={args['src']}
 	>
 		{args['children']}
@@ -57,16 +63,19 @@ boolean('captionOverlayOnFocus', _Card, Config);
 boolean('centered', _Card, Config);
 boolean('disabled', _Card, Config);
 boolean('fitImage', _Card, Config);
+select('icon', _Card, iconsList, Config);
 object('imageIconSrc', _Card, Config, generateImageSrc('0084ff'));
 object('imageSize', _Card, Config);
 boolean('hasContainer', _Card, Config);
 text('label', _Card, Config, 'Card label');
 select('orientation', _Card, prop.orientation, Config);
 object('primaryBadgeSrc', _Card, Config, generateImageSrc('ff6d78'));
+number('progress', _Card, Config, 0.5);
 boolean('roundedImage', _Card, Config);
 object('secondaryBadgeSrc', _Card, Config, generateImageSrc('ffc600'));
 text('secondaryLabel', _Card, Config, 'Card secondary label');
 boolean('selected', _Card, Config);
+boolean('showProgressBar', _Card, Config);
 object('src', _Card, Config, generateImageSrc('93d371'));
 text('children', _Card, Config, 'Card Caption');