+++ tests/screenshot/apps/components/Card.js
@@ -53,6 +53,12 @@ const CardTests = [
 	// Centered
 	...withProps({centered: true}, defaultCardTests),
 
+	// Icon
+	...withProps({icon: 'trash'}, defaultCardTests),
+
+	// Show ProgressBar
+	...withProps({progress: 0.5, showProgressBar: true}, defaultCardTests),
+
 	// Focused
 	...withConfig({focus: true, wrapper: {padded: true}}, defaultCardTests),